feat(website): static-render solution pages with per-problem metadata
Solution pages were fully client-rendered, so all 269 routes were dynamic
and shared the root layout's generic title. Split into a server page.tsx +
ProblemPageClient, mirroring app/forum/[id].

- generateStaticParams derives routes from problems[].link, deduped (289
  links resolve to 269 routes since shared junior problems point at the
  senior page)
- generateMetadata emits a unique title/description plus canonical, OG and
  twitter tags from the problem's name, difficulty and tags

Also fix the forum post description truncating at 100 chars against a
155-char threshold, memoize getPost with react cache() to stop the double
PocketBase call per request, and drop the GitHub star count to a 24h
revalidate.
